Add IP_TRANSPARENT option
The IP_TRANSPARENT socket option is required for our short-term approach to
triangle routing. In this approach, we plan to run large proxy processes, each
serving many public IPs. We will use the iptables TPROXY target to redirect
packets, destined for each public IP, to the proxy process. In order for the
process to receive these packets, it must listen on a socket with
IP_TRANSPARENT.