Prohibit localhost connections under DNS rebinding attacks (#606)
* Prohibit localhost connections under DNS rebinding attacks

We make 2 IP resolution calls when a connection is made. This opens us
up for a DNS rebinding attack where the second call was susceptible to
DNS resolving to a local address on our network. This change ensures
that we make a check of the resolved IP address on both the first and
second call